requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
joblib==0.13.2
nltk==3.4.3
numpy==1.16.4
praw==6.2.0
prawcore==1.0.1
python-decouple==3.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
update-checker==0.16
urllib3==1.25.3
websocket-client==0.56.0
